fix(scan): use Double for public rating scores and support half stars
Public ratings from Gemini use decimal scores (e.g. Goodreads 4.18/5)
which caused iOS DecodingError with Int types. StarRatingView now
supports half-star display with star.leadinghalf.filled. Personal
ratings remain Int (full stars only).

+++ ios/Pchook/Shared/Book.swift
@@ -37,8 +37,8 @@ struct Award: Codable, Sendable {
 
 struct PublicRating: Codable, Sendable {
     let source: String
-    let score: Int
-    let maxScore: Int
+    let score: Double
+    let maxScore: Double
     let voterCount: Int
 }
